Implemented more physics systems and properties
Constant for toxin projectile physics sphere size

Bunch more work on the toxin entity, C# side is almost entirely done

just need a bit more of the collision stuff done

Made physics collision management compile for now

still missing the logic to go from physics bodies to entity references

Partially converted the floating chunk entity to ECS

Remove trying to compile the task system file

which was not anything yet

Converted two more systems needed by chunks

More work on the ecs conversion

Finished converting floating chunk

all that's left are things with "microbe" in the name

Commenting out some stuff to get closer to compiling code

+++ src/engine/common_components/Spawned.cs
@@ -6,10 +6,9 @@
     public struct Spawned
     {
         /// <summary>
-        ///   If the squared distance to the player of this object is
-        ///   greater than this, it is despawned.
+        ///   If the squared distance to the player of this object is greater than this, it is despawned.
         /// </summary>
-        public int DespawnRadiusSquared;
+        public float DespawnRadiusSquared;
 
         /// <summary>
         ///   How much this entity contributes to the entity limit relative to a single node
